+++ perf/conformation/dwt/dwt.h
@@ -1,0 +1,9 @@
+#ifndef __CONFORMATION_DWT_H__
+#define __CONFORMATION_DWT_H__
+
+#include <ffitt/transform/dwt.h>
+#include <perf/conformation/support.h>
+
+void run_dwt_conformation_tests(void);
+
+#endif//__CONFORMATION_DWT_H__