+++ perf/conformation/eigen/eigen.h
@@ -1,0 +1,9 @@
+#ifndef __CONFORMATION_EIGEN_H__
+#define __CONFORMATION_EIGEN_H__
+
+#include <ffitt/linalg/eigen.h>
+#include <perf/conformation/support.h>
+
+void run_eigen_conformation_tests(void);
+
+#endif//__CONFORMATION_EIGEN_H__