+++ perf/conformation/poly/poly.h
@@ -1,0 +1,9 @@
+#ifndef __CONFORMATION_POLY_H__
+#define __CONFORMATION_POLY_H__
+
+#include <ffitt/linalg/poly.h>
+#include <perf/conformation/support.h>
+
+void run_poly_conformation_tests(void);
+
+#endif//__CONFORMATION_POLY_H__