+++ perf/conformation/vector/vector.c
@@ -4,7 +4,7 @@
 static void run_vector_dot_product_static_conformation(void)
 {
     vector_t A0, B0;
-    float dot_product0;
+    real_t dot_product0;
     VECTOR_INIT(A0, 3, ((float[]){1, 2, 3}));
     VECTOR_INIT(B0, 3, ((float[]){4, 5, 6}));
     dot_product0 = vector_dot_product(&A0, &B0);
@@ -13,7 +13,7 @@ static void run_vector_dot_product_static_conformation(void)
     VECTOR_FREE(B0);
 
     vector_t A1, B1;
-    float dot_product1;
+    real_t dot_product1;
     VECTOR_INIT(A1, 4, ((float[]){-1.5f, 2.5f, -3.5f, 4.5f}));
     VECTOR_INIT(B1, 4, ((float[]){-4.5f, 3.5f, -2.5f, 1.5f}));
     dot_product1 = vector_dot_product(&A1, &B1);
@@ -22,7 +22,7 @@ static void run_vector_dot_product_static_conformation(void)
     VECTOR_FREE(B1);
 
     vector_t A2, B2;
-    float dot_product2;
+    real_t dot_product2;
     VECTOR_INIT(A2, 2, ((float[]){7.5f, -8.5f}));
     VECTOR_INIT(B2, 2, ((float[]){-9.5f, 10.5f}));
     dot_product2 = vector_dot_product(&A2, &B2);
@@ -43,14 +43,14 @@ static void run_vector_dot_product_dynamic_conformation(void)
 static void run_vector_norm_static_conformation(void)
 {
     vector_t A0;
-    float norm0;
+    real_t norm0;
     VECTOR_INIT(A0, 3, ((float[]){1, 2, 2}));
     norm0 = vector_norm(&A0);
     VALUE_CHECK_EQUAL_CASE(norm0, 3.0f, "Vector Static Norm Test");
     VECTOR_FREE(A0);
 
     vector_t A1;
-    float norm1;
+    real_t norm1;
     VECTOR_INIT(A1, 4, ((float[]){-1.0f, 1.0f, -1.0f, 1.0f}));
     norm1 = vector_norm(&A1);
     VALUE_CHECK_EQUAL_CASE(norm1, 2.0f, "Vector Static Norm Test");